Struct metrics_distributor::collectors::http::LogDrainHandler

pub struct LogDrainHandler {
    // some fields omitted
}

Accepts HTTP requests and reads lines from the body. Each line will be passed to its set of readers and any metrics collected by those readers will be recorded in the store.

Methods

impl LogDrainHandler

fn new(store: SharedStore, readers: Vec<Box<LogLineReader>>) -> LogDrainHandler

Create a new handler with the given vector of readers. Each reader will be called on every line the handler receives.