Struct metrics_distributor::metrics::AggregatedMetrics

pub struct AggregatedMetrics {
    // some fields omitted
}

All the metrics in a given time interval coalesced into a single value for each metric.

Methods

impl AggregatedMetrics

fn new() -> AggregatedMetrics

fn with_metrics(metrics: Vec<AggregatedMetric>) -> AggregatedMetrics

fn aggregate_counts(&mut self, counts: Iter<String, u64>)

fn aggregate_measures(&mut self, measures: Iter<String, Vec<f64>>)

Rolls up all the given measures. The minimum, maximum, median, average (mean), and 95th percentile summary measures will all be emitted, as well as a total count of all the individual measures received in the period.

fn aggregate_samples(&mut self, samples: Iter<String, f64>)

fn iter(&self) -> Iter<AggregatedMetric>

fn len(&self) -> usize